+++ release-notes.md
@@ -2,6 +2,8 @@
 
 ### New Features
 
+- **Notifications** - Add a notification toast when new songs are played.
+
 ### Bug Fixes
 
 - **Remove console** —  Remove windows console to show when opening the application.